Refactor get_type_hints logic
